feat(blog): move components to root folder
