+++ pages/_app.js
@@ -6,6 +6,4 @@ function MyApp({ Component, pageProps }) {
 	return <Component {...pageProps} />;
 }
 
-console.log('dale', process.env.GOOGLE_ANALYTICS);
-
 export default withGA(process.env.GOOGLE_ANALYTICS, Router)(MyApp);
