+++ pages/about.js
@@ -1,4 +1,4 @@
-import Layout from './components/layout/';
+import Layout from '../components/layout/';
 
 export default function About() {
 	return (
