+++ pages/blog/p/goodbye-clean-code.mdx
@@ -5,7 +5,7 @@ export const frontmatter = {
   heroImage: '../static/bali-15.jpg'
 };
 
-import Layout from '../../components/blog';
+import Layout from '../../../components/blog';
 
 export default ({ children }) => <Layout>{children}</Layout>;
