+++ pages/blog/p/otherpost.mdx
@@ -1,4 +1,4 @@
-import Layout from '../../components/blog';
+import Layout from '../../../components/blog';
 
 export default ({ children }) => <Layout>{children}</Layout>
