--- pages/components/code-block/index.jsx
@@ -1,21 +1,0 @@
-import React from 'react';
-import Highlight, { defaultProps } from 'prism-react-renderer';
-
-export default ({ children, className }) => {
-  const language = className.replace(/language-/, '');
-  return (
-    <Highlight {...defaultProps} code={children} language={language}>
-      {({ className, style, tokens, getLineProps, getTokenProps }) => (
-        <pre className={`text-l rounded-lg m-4 ${className}`} style={{ ...style, padding: '20px' }}>
-          {tokens.map((line, i) => (
-            <div key={i} {...getLineProps({ line, key: i })}>
-              {line.map((token, key) => (
-                <span key={key} {...getTokenProps({ token, key })} />
-              ))}
-            </div>
-          ))}
-        </pre>
-      )}
-    </Highlight>
-  );
-};
